Split ResolverFactory into a ComponentResolversFactory and DependencyGraphResolvers
These were two separate concepts stuck in a single file.
Separating them lets us avoid non-static inner classes
